Minor fix to executeBatch
Currently the error message reported by `executeBatch` method is
incorrect because this method tries to clear the batch on exit and such
clearance is not possible when the statement is already closed after
the error. This change adds a check that the statement is open before
doing the cleanup.

Testing: existing test is enhanced to check this error message; batch
tests are moved to separate file.
